GCFloat doesn't seem to be available during command-line build (??)

+++ src/libambulant/gui/cocoa/cocoa_ink.mm
@@ -87,7 +87,7 @@ cocoa_ink_renderer::cocoa_ink_renderer(
 	m_color = [NSColor colorWithCalibratedRed:redf(color)
 					green:greenf(color)
 					blue:bluef(color)
-					alpha:(CGFloat)1.0];
+					alpha:1.0f];
 }
 
 cocoa_ink_renderer::~cocoa_ink_renderer()
@@ -153,7 +153,7 @@ cocoa_ink_renderer::redraw_body(const rect &dirty, gui_window *window)
 				double x = strtod(x_str, NULL);
 				double y = strtod(y_str, NULL);
 				AM_DBG lib::logger::get_logger()->debug("cocoa_ink_renderer:     point (%f,%f)", x, y);
-				NSPoint pt = NSMakePoint((CGFloat)x, (CGFloat)y);
+				NSPoint pt = NSMakePoint((float)x, (float)y);
 				if (start_new_trace) {
 					[m_path moveToPoint: pt];
 					start_new_trace = false;
