+++ src/libambulant/gui/cocoa/cocoa_transition.mm
@@ -62,7 +62,7 @@ finalize_transition_bitblit(bool outtrans, common::surface *dst)
 		[[view getTransitionNewSource] drawInRect: cocoa_dstrect_whole 
 			fromRect: cocoa_dstrect_whole
 			operation: NSCompositeSourceOver
-			fraction: (CGFloat)1.0];
+			fraction: 1.0f];
 	}
 }
 
@@ -79,7 +79,7 @@ cocoa_transition_blitclass_fade::update()
 	[newsrc drawInRect: cocoa_dstrect_whole 
 		fromRect: cocoa_dstrect_whole
 		operation: NSCompositeSourceOver
-		fraction: (CGFloat)m_progress];
+		fraction: (float)m_progress];
 	finalize_transition_bitblit(m_outtrans, m_dst);
 }
 
@@ -98,7 +98,7 @@ cocoa_transition_blitclass_rect::update()
 	[newsrc drawInRect: cocoa_newrect_whole 
 		fromRect: cocoa_newrect_whole
 		operation: NSCompositeSourceOver
-		fraction: (CGFloat)1.0];
+		fraction: 1.0f];
 	finalize_transition_bitblit(m_outtrans, m_dst);
 }
 
@@ -131,22 +131,22 @@ cocoa_transition_blitclass_r1r2r3r4::update()
 		[newsrc drawInRect: cocoa_olddstrect_whole 
 			fromRect: cocoa_oldsrcrect_whole
 			operation: NSCompositeCopy
-			fraction: (CGFloat)1.0];
+			fraction: 1.0f];
 
 		[oldsrc drawInRect: cocoa_newdstrect_whole 
 			fromRect: cocoa_newsrcrect_whole
 			operation: NSCompositeSourceOver
-			fraction: (CGFloat)1.0];
+			fraction: 1.0f];
 	} else {
 		[oldsrc drawInRect: cocoa_olddstrect_whole 
 			fromRect: cocoa_oldsrcrect_whole
 			operation: NSCompositeCopy
-			fraction: (CGFloat)1.0];
+			fraction: 1.0f];
 
 		[newsrc drawInRect: cocoa_newdstrect_whole 
 			fromRect: cocoa_newsrcrect_whole
 			operation: NSCompositeSourceOver
-			fraction: (CGFloat)1.0];
+			fraction: 1.0f];
 	}
 }
 
@@ -167,7 +167,7 @@ cocoa_transition_blitclass_rectlist::update()
 		[newsrc drawInRect: cocoa_newrect_whole 
 			fromRect: cocoa_newrect_whole
 			operation: NSCompositeSourceOver
-			fraction: (CGFloat)1.0];
+			fraction: 1.0f];
 	}
 	finalize_transition_bitblit(m_outtrans, m_dst);
 }
@@ -213,11 +213,11 @@ composite_path(AmbulantView *view, lib::rect dstrect_whole, NSBezierPath *path, 
 		newalpha = 1.0F;
 	}
 	[tmpsrc lockFocus];
-	[[NSColor colorWithDeviceWhite: (CGFloat)1.0 alpha: (CGFloat)oldalpha] set];
+	[[NSColor colorWithDeviceWhite: 1.0f alpha: (float)oldalpha] set];
 	NSRectFill(cocoa_dstrect_whole);
 
 	// Now we fill draw the path on the temp bitmap, with opaque white
-	[[NSColor colorWithDeviceWhite: (CGFloat)1.0 alpha: (CGFloat)newalpha] set];
+	[[NSColor colorWithDeviceWhite: 1.0f alpha: (float)newalpha] set];
 	[path fill];
 
 	// Next we composit the source image onto the temp bitmap, but only where
@@ -225,7 +225,7 @@ composite_path(AmbulantView *view, lib::rect dstrect_whole, NSBezierPath *path, 
 	[newsrc drawInRect: cocoa_dstrect_whole 
 		fromRect: cocoa_dstrect_whole
 		operation: NSCompositeSourceIn
-		fraction: (CGFloat)1.0];
+		fraction: 1.0f];
 
 	// Finally we put the opaque bits of the temp image onto the destination
 	// image
@@ -233,7 +233,7 @@ composite_path(AmbulantView *view, lib::rect dstrect_whole, NSBezierPath *path, 
 	[tmpsrc drawInRect: cocoa_dstrect_whole 
 		fromRect: cocoa_dstrect_whole
 		operation: NSCompositeSourceOver
-		fraction: (CGFloat)1.0];
+		fraction: 1.0f];
 }
 
 void
